+++ components/Banner/ColorText.tsx
@@ -8,7 +8,7 @@ interface ColorTextProps {
   onClick?: VoidFunction;
 }
 
-export function ColorText(props: ColorTextProps): React.ReactElement {
+function ColorText(props: ColorTextProps): React.ReactElement {
   const { className, text, backgroundColor, url, onClick } = props;
 
   return (
@@ -28,3 +28,5 @@ export function ColorText(props: ColorTextProps): React.ReactElement {
     </a>
   );
 }
+
+export default ColorText;
