+++ components/Layout.tsx
@@ -4,7 +4,7 @@ interface LayoutProps {
   children: React.ReactNode;
 }
 
-export function Layout(props: LayoutProps): React.ReactElement {
+function Layout(props: LayoutProps): React.ReactElement {
   const { children } = props;
 
   return (
@@ -16,3 +16,5 @@ export function Layout(props: LayoutProps): React.ReactElement {
     </div>
   );
 }
+
+export default Layout;
