--- components/List/index.ts
@@ -1,2 +1,0 @@
-export * from './List';
-export * from './TagList';
