+++ components/TagList/Stack.tsx
@@ -1,0 +1,26 @@
+import { Tag } from 'antd';
+import React, { useCallback } from 'react';
+import { Stack, StackInfo } from '../../config';
+import BaseTagList from './Base';
+
+interface StackListProps {
+  stack: Stack[];
+}
+
+function StackList(props: StackListProps): React.ReactElement {
+  const { stack } = props;
+
+  const renderStack = useCallback((stack: Stack): React.ReactNode => {
+    const { color, value } = StackInfo[stack];
+
+    return (
+      <div className='pb-1'>
+        <Tag color={color}>{value}</Tag>
+      </div>
+    );
+  }, []);
+
+  return <BaseTagList title='Skills' data={stack} renderList={renderStack} />;
+}
+
+export default StackList;
