--- components/index.ts
@@ -1,10 +1,0 @@
-export * from './Banner';
-export * from './BottomSheet';
-export * from './Card';
-export * from './Conditional';
-export * from './Footer';
-export * from './Header';
-export * from './Layout';
-export * from './List';
-export * from './SubHeader';
-export * from './Seo';
